// src/app/core/models/diet.ts
export class Diet {
  id: number;
  name: string;

  mondayEarlyMorning: string;
  mondayBreakfast: string;
  mondayMidMorning: string;
  mondayLunch: string;
  mondayMidNoon: string;
  mondayDinner: string;

  tuesdayEarlyMorning: string;
  tuesdayBreakfast: string;
  tuesdayMidMorning: string;
  tuesdayLunch: string;
  tuesdayMidNoon: string;
  tuesdayDinner: string;

  wednesdayEarlyMorning: string;
  wednesdayBreakfast: string;
  wednesdayMidMorning: string;
  wednesdayLunch: string;
  wednesdayMidNoon: string;
  wednesdayDinner: string;

  thursdayEarlyMorning: string;
  thursdayBreakfast: string;
  thursdayMidMorning: string;
  thursdayLunch: string;
  thursdayMidNoon: string;
  thursdayDinner: string;

  fridayEarlyMorning: string;
  fridayBreakfast: string;
  fridayMidMorning: string;
  fridayLunch: string;
  fridayMidNoon: string;
  fridayDinner: string;

  saturdayEarlyMorning: string;
  saturdayBreakfast: string;
  saturdayMidMorning: string;
  saturdayLunch: string;
  saturdayMidNoon: string;
  saturdayDinner: string;

  sundayEarlyMorning: string;
  sundayBreakfast: string;
  sundayMidMorning: string;
  sundayLunch: string;
  sundayMidNoon: string;
  sundayDinner: string;
}